chore: Remove `deleted` field on domain classes (#29900)
Use `deletedAt` only hereon.

We're not changing the `notDeleted()` query function, we still check for
boolean deleted there, for two reasons.

1. There may be objects that have `deleted: true`, but no `deletedAt`.
Safer to just include that in the query, than to migrate now.
2. A lot of indexes are built taking both these fields into
consideration. We don't know how the query without the boolean field
would perform with such indexes. Recreating indexes now, is also a
time-consuming migration.

Depends on #29924.