chore: Fluent interface for queryAll method (#30758)
There is ~13 signatures of the `queryAll` method with parameters ranging
from 2 to 7, different combinations of the same superset. This is
extremely hard to grok and the resulting uses of `queryAll` aren't very
readable either. This PR introduces a fluent interface for this method.
If this works better for us, I'll look to move more such
over-signature-ed methods to a fluent interface.

This sprang from a frustration in trying to understand this method's
usages, to port to Postgres, and pulling my hair apart in the process.

Here's an example of the fluent API compared to current API:

![shot-2024-01-27-10-37-57@2x](https://github.com/appsmithorg/appsmith/assets/120119/ed11a4c6-a433-4e92-824c-7db00880a24a)