test: Add AssertJ for plugins, for better assertions (#31380)
PR adds assertj to the plugin modules as well so we can write better
assertions there. We're changing just one test, but to illustrate as an
example.

With the previous assertion used here, failure messages looked like
this:

```
Expected <true>, but got <false>.
```

But with the better assertions introduced in this PR, we should see
something like:

```
Expected `Random unexpected string` to end with `Host not allowed.`
```

Which should help us much better in troubleshooting.