chore: Remove BaseController on UserControllerCE (#32216)
The `BaseController` is used by only five controller classes. Regarding
the `GET /` route in the base controller,

- two override and block it (`Application` and `User`).
- two override with a custom implementation and logic, completely
ignoring the `params` object (`Theme` and `Workspace`).
- one appears to be using it (`Plugin`).

This makes it confusing and hard-to-maintain. This common route is
overridden more times than it's reused.

This PR removes the `BaseController` on `UserController` as step 1. We
intend to remove it from the remaining 3 in subsequent PRs.

Server and Cypress tests verified on EE.